Fixed docstring typo.
