+++ diags/grids/bearing_geoloc.py
@@ -32,7 +32,7 @@ Requirements
 
 - astropy; https://github.com/astropy/astropy
 
-- ufs_pytils; https://github.com/HenryWinterbottom-NOAA/ufs_pyutils
+- ufs_pyutils; https://github.com/HenryWinterbottom-NOAA/ufs_pyutils
 
 Author(s)
 ---------
