+++ diags/interp/ll2ra.py
@@ -1,20 +1,9 @@
 # =========================================================================
-
-# Module: interp/ll2ra.py
-
+# File: diags/interp/ll2ra.py
 # Author: Henry R. Winterbottom
-
-# Email: henry.winterbottom@noaa.gov
-
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the respective public license published by the
-# Free Software Foundation and included with the repository within
-# which this application is contained.
-
-# This program is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-
+# Date: 07 March 2023
+# Version: 0.0.1
+# License: LGPL v2.1
 # =========================================================================
 
 """
@@ -42,7 +31,7 @@ Functions
 Requirements
 ------------
 
-- ufs_pytils; https://github.com/HenryWinterbottom-NOAA/ufs_pyutils
+- ufs_pyutils; https://github.com/HenryWinterbottom-NOAA/ufs_pyutils
 
 Author(s)
 ---------
@@ -73,13 +62,9 @@ __email__ = "henry.winterbottom@noaa.gov"
 from types import SimpleNamespace
 
 import numpy
-from grids.haversine import haversine
+from diags.grids.haversine import haversine
+from scipy.interpolate import griddata
 from tools import parser_interface
-from utils.logger_interface import Logger
-
-# ----
-
-logger = Logger(caller_name=__name__)
 
 # ----
 
@@ -171,62 +156,26 @@ def ll2ra(
     varobj.dphi = numpy.radians(dphi)
     varobj.drho = drho
     varobj.max_radius = max_radius
+    varobj.radial = numpy.arange(0.0, (varobj.max_radius + varobj.drho), varobj.drho)
+    varobj.azimuth = numpy.arange(
+        -1.0 * numpy.pi, (numpy.pi + 2.0 * varobj.dphi), varobj.dphi
+    )
 
     # Compute the radial distance relative to the specified
     # geographical coordinate location.
     rho = numpy.array(
         [haversine(fix, (lats[idx], lons[idx])) for idx in range(len(lats))]
     )
-    xx = numpy.array([haversine(fix, (lat_0, lons[idx]))
-                     for idx in range(len(lats))])
+    xx = numpy.array([haversine(fix, (lat_0, lons[idx])) for idx in range(len(lats))])
     xx = numpy.where(lons < lon_0, -1.0 * xx, xx)
-
-    yy = numpy.array([haversine(fix, (lats[idx], lon_0))
-                     for idx in range(len(lats))])
+    yy = numpy.array([haversine(fix, (lats[idx], lon_0)) for idx in range(len(lats))])
     yy = numpy.where(lats < lat_0, -1.0 * yy, yy)
     phi = numpy.arctan2(yy, xx)
 
-    # Interpolate the Cartesian grid to the defined polar coordinate
-    # grid.
-    varobj.radial = numpy.arange(
-        0.0, (varobj.max_radius + varobj.drho), varobj.drho)
-    varobj.azimuth = numpy.arange(
-        (-1.0 * numpy.pi), (numpy.pi + varobj.dphi), varobj.dphi
-    )
-    msg = (
-        f"Defining polar projection grid of resolution {varobj.drho} meters "
-        f"and {varobj.dphi} radians centered at longitude coordinate {varobj.lon_0} "
-        f"and latitude coordinate location {varobj.lat_0}."
-    )
-    logger.info(msg=msg)
-
-    # Interpolate the variable defined on the Cartesian grid to the
-    # established the radial and azimuthal angle coordinates; proceed
-    # accordingly.
-    var = []
-
-    # Define the radial coordinate.
-    for ridx in enumerate(varobj.radial):
-        radii = varobj.radial[ridx[0]]
-
-        # Define the azimuthal coordinate.
-        for aidx in enumerate(varobj.azimuth):
-            theta = varobj.azimuth[aidx[0]]
-            array = []
-
-            # Determine all (if any) input variable values within the
-            # radii and azimuth (theta) interval; proceed accordingly.
-            idxs = numpy.ndarray.tolist(numpy.where(rho >= radii)[0])
-            for idx in idxs:
-                if rho[idx] < (radii + varobj.drho):
-                    if (phi[idx] >= theta) and (phi[idx] < (theta + varobj.dphi)):
-                        array.append(varin[idx])
-            if len(array) == 0:
-                var.append(numpy.nan)
-            else:
-                var.append(numpy.nanmean(array))
-
-    # Interpolate in order to fill and missing (i.e., NaN) values.
+    # Interpolate the variable from the Cartesian projection to the
+    # polar projection.
+    (r_mesh, theta_mesh) = numpy.meshgrid(varobj.radial, varobj.azimuth)
+    var = griddata((rho, phi), varin, (r_mesh, theta_mesh), method="linear").ravel()
     interp_var = numpy.array(var)
     check = numpy.logical_not(numpy.isnan(interp_var))
     xp = check.ravel().nonzero()[0]
@@ -234,6 +183,6 @@ def ll2ra(
     x = numpy.isnan(interp_var).ravel().nonzero()[0]
     interp_var[numpy.isnan(var)] = numpy.interp(x, xp, fp)
     (varobj.nrho, varobj.nphi) = [len(varobj.radial), len(varobj.azimuth)]
-    varobj.varout = numpy.array(interp_var).reshape((varobj.nrho, varobj.nphi))
+    varobj.varout = numpy.array(interp_var).reshape((varobj.nphi, varobj.nrho)).T
 
     return varobj
